fix: Handle `6k + 1` candidates properly
* Before this change, calling `find_next_prime` with a candidate which
  satisfies `6k + 1` pattern did not work. This pattern was only
  later in the loop, but not for the original `n` argument.
* Use `u32` as a type for `n`. It's safe to convert `u32` into `f64`
  without precission loss.

+++ xtask/src/hash_set.rs
@@ -59,7 +59,7 @@ fn bench(opts: BenchOptions) -> anyhow::Result<()> {
 
     for load_factor in opts.load_factors {
         let capacity_with_load_factor =
-            find_next_prime_with_load_factor(opts.expected_capacity as f64, load_factor) as usize;
+            find_next_prime_with_load_factor(opts.expected_capacity, load_factor) as usize;
 
         let mut hs = HashSet::new(capacity_with_load_factor, opts.sequence_threshold)?;
         let mut rng = thread_rng();
